Next Greater Element using Stack
In this problem, you have to implement int[] nextGreaterElement(int[] arr) method. For each element in an array, it finds the next greater element in that array.o keep it simple, the next greater element for the last or maximum value in the array is -1.In each iteration, we only check the array elements appearing after the current element.