Check for Balanced Parentheses using a Stack
In this problem, you have to implement the isBalanced() method, which will take a string containing only curly {}, square [], and round () parentheses. The method will tell us whether all the parentheses in the string are balanced or not.

For all the parentheses to be balanced, every opening parenthesis must have a closing one. The order in which they appear, also matters. For example, {[]} is balanced, but {[}] is not.